+++ src/components/layout.js
@@ -50,4 +50,4 @@ const Layout = ({ location, title, children }) => {
   )
 }
 
-export default Layout
+export default Layout;
