+++ src/pages/404.js
@@ -15,6 +15,8 @@ const NotFoundPage = ({ data, location }) => {
   )
 }
 
+export const Head = () => <Seo title="404: Not Found" />
+
 export default NotFoundPage
 
 export const pageQuery = graphql`
