+++ src/style.css
@@ -47,7 +47,7 @@
   --fontSize-5: 2.074rem;
   --fontSize-6: 2.488rem;
   --fontSize-7: 2.986rem;
-  --color-primary: #005b99;
+  --color-primary: #FAFAFA;
   --a-color-primary: #22c1c3;
   --color-text: #e6e6e6;
   --color-text-light: #4f5969;
