+++ src/templates/tutorial-post.js
@@ -3,7 +3,6 @@ import { Link, graphql } from "gatsby"
 
 import Bio from "../components/bio"
 import Layout from "../components/layout"
-import Seo from "../components/seo"
 
 const TutorialPostTemplate = ({
   data: { previous, next, site, markdownRemark: post },
